fix(tui-rs): fix CI compatibility issues with mlua and wasmtime
## mlua fixes
- Added "send" feature for thread-safety (Arc instead of Rc)
- Added "vendored" feature to build Lua from source
- Map mlua errors to anyhow to satisfy Send bounds
- Import LuaSerdeExt for serialize methods
- Fix sandbox code (globals() already returns Table)

## wasmtime fixes
- Use reborrows (&mut *store) for multiple calls

## Clippy fixes
- Add #[allow(dead_code)] for unused structs
- Add #[allow(clippy::should_implement_trait)] for from_str
- Use strip_prefix instead of manual slice
- Add type alias for complex handler type

All tests and clippy checks pass with --all-features.

+++ packages/tui-rs/src/hooks/integration.rs
@@ -54,6 +54,7 @@ pub struct IntegratedHookSystem {
 
 /// TypeScript hook configuration
 #[derive(Debug, Clone)]
+#[allow(dead_code)]
 struct TypeScriptHookConfig {
     path: PathBuf,
     event: HookEventType,
@@ -89,7 +90,7 @@ impl IntegratedHookSystem {
         match load_hook_config(Path::new(cwd)) {
             Ok(config) => {
                 system.enabled = config.settings.enabled;
-                system.timeout = Duration::from_millis(config.settings.timeout_ms as u64);
+                system.timeout = Duration::from_millis(config.settings.timeout_ms);
                 system.log_file = config.settings.log_file.clone();
                 system.load_hooks_from_config(&config);
 
@@ -198,7 +199,7 @@ impl IntegratedHookSystem {
         // Reload config
         if let Ok(config) = load_hook_config(Path::new(&self.cwd)) {
             self.enabled = config.settings.enabled;
-            self.timeout = Duration::from_millis(config.settings.timeout_ms as u64);
+            self.timeout = Duration::from_millis(config.settings.timeout_ms);
             self.log_file = config.settings.log_file.clone();
         }
